+++ .snyk
@@ -3,10 +3,10 @@ version: v1.7.0
 # ignores vulnerabilities until expiry date; change duration by modifying expiry date
 ignore:
   SNYK-RUBY-NOKOGIRI-20299:
-    - '* > nokogiri@1.7.1':
-        reason: None given
-        expires: '2017-06-09T16:51:02.975Z'
     - '* > nokogiri@1.7.2':
         reason: None given
         expires: '2017-06-21T12:12:22.642Z'
+    - '* > nokogiri@1.8.2':
+        reason: None given
+        expires: '2018-03-14T12:12:22.642Z'
 patch: {}
