+++ app/controllers/api/v1/sections_controller.rb
@@ -13,6 +13,10 @@ module Api
 
         respond_with @section
       end
+
+      def tree
+        @sections = Section.eager({ chapters: [:headings] }).all
+      end
     end
   end
 end
