+++ app/views/api/v1/headings/show.json.rabl
@@ -4,6 +4,12 @@ cache @heading_cache_key, expires_at: actual_date.end_of_day
 attributes :goods_nomenclature_item_id, :description, :bti_url,
            :formatted_description
 
+unless @heading.footnote.nil?
+  child :footnote do
+    attributes :code, :description, :formatted_description
+  end           
+end  
+
 if @heading.declarable?
   attributes :basic_duty_rate
