+++ app/views/api/v1/headings/tree.json.rabl
@@ -1,0 +1,6 @@
+object @heading
+attributes :id, :declarable
+
+child(@commodities) do
+    attributes :id
+end
