+++ app/views/api/v1/sections/tree.json.rabl
@@ -1,0 +1,10 @@
+collection @sections
+attributes :id
+
+child(chapters: :chapters) do
+  attributes :goods_nomenclature_item_id
+  child(headings: :headings) do
+    attributes :goods_nomenclature_item_id,
+               :declarable
+  end
+end
