+++ lib/tasks/tariff.rake
@@ -12,6 +12,42 @@ namespace :tariff do
     TradeTariffBackend.reindex
   end
 
+  desc 'Add commodity footnotes for ECO licences where these is an export restriction'
+  task add_missing_commodity_footnote: :environment do
+    measure_type_id = MeasureType.all.detect { |mt| mt.description == 'Export authorization (Dual use)' }.values[:measure_type_id]
+    footnote = Footnote.filter(footnote_id: '002', footnote_type_id: '05').first
+    sql = %Q{
+      SELECT DISTINCT goods_nomenclatures.goods_nomenclature_sid, goods_nomenclatures.goods_nomenclature_item_id,
+      goods_nomenclatures.validity_start_date, producline_suffix,
+      goods_nomenclatures.validity_end_date, goods_nomenclatures.operation
+      FROM goods_nomenclatures
+      LEFT JOIN measures ON (measures.goods_nomenclature_item_id = goods_nomenclatures.goods_nomenclature_item_id)
+         WHERE ((measures.measure_type_id = '#{measure_type_id}')
+        )
+      AND NOT EXISTS (
+      select 1 from footnote_association_goods_nomenclatures
+      where footnote_association_goods_nomenclatures.footnote_type = '05'
+      and footnote_association_goods_nomenclatures.footnote_id = '002'
+      and goods_nomenclature_item_id = goods_nomenclatures.goods_nomenclature_item_id
+      )
+    }
+    GoodsNomenclature.fetch(sql).all.each do |c|
+      if c.footnote != footnote
+        f = FootnoteAssociationGoodsNomenclature.new
+        f.values[:goods_nomenclature_sid] = c.values[:goods_nomenclature_sid]
+        f.values[:goods_nomenclature_item_id] = c.values[:goods_nomenclature_item_id]
+        f.values[:productline_suffix] = c.values[:producline_suffix]
+        f.values[:validity_start_date] = c.values[:validity_start_date]
+        f.values[:validity_end_date] = c.values[:validity_end_date]
+        f.values[:operation] = c.values[:operation]
+        f.values[:footnote_id] = footnote.values[:footnote_id]
+        f.values[:footnote_type] = footnote.values[:footnote_type_id]
+        f.values[:national] = footnote.values[:national]
+        f.save
+      end
+    end  
+  end
+
   desc 'Download and apply Taric and CHIEF data'
   task sync: %w[environment sync:apply]
